requirements.txt
fastapi
uvicorn
lightgbm
lightkurve
astropy
numpy
pandas
scikit-learn
pyyaml
pyarrow
python-multipart
